@@ -3,11 +3,20 @@ package databaser
 
 import (
 	"context"
+	"database/sql"
+	_ "embed"
+	"errors"
 	"fmt"
+	"log/slog"
+	"strconv"
+	"time"
 
 	"github.com/jmoiron/sqlx"
 	_ "modernc.org/sqlite" // SQLite driver
 )
+
+//go:embed init.sql
+var initSQL string
 
 // DB wraps sqlx.DB for database operations.
 type DB struct {
@@ -15,73 +24,171 @@ type DB struct {
 }
 
 // New creates a new database connection.
-func New(path string) (*DB, error) {
+func New(ctx context.Context, path string) (*DB, error) {
 	db, err := sqlx.Open("sqlite", path)
 	if err != nil {
 		return nil, fmt.Errorf("open database: %w", err)
 	}
 
-	// Set connection pool settings for SQLite
+	pragmas := []string{
+		"PRAGMA journal_mode=WAL",   // write-ahead logging
+		"PRAGMA synchronous=NORMAL", // balance between performance and safety
+		"PRAGMA cache_size=-32000",  // 32 mb cache
+		"PRAGMA busy_timeout=5000",  // 5 сек cache wait
+		"PRAGMA foreign_keys=ON",    // enable foreign key constraints
+	}
+
+	for _, pragma := range pragmas {
+		if _, err = db.ExecContext(ctx, pragma); err != nil {
+			return nil, fmt.Errorf("set pragma %q: %w", pragma, err)
+		}
+	}
+
 	db.SetMaxOpenConns(1) // SQLite doesn't support multiple writers
 
-	if err := db.Ping(); err != nil {
-		db.Close()
+	if err = db.Ping(); err != nil {
+		if closeErr := db.Close(); closeErr != nil {
+			slog.ErrorContext(ctx, "failed to close database after ping error", "error", closeErr)
+		}
 		return nil, fmt.Errorf("ping database: %w", err)
 	}
 
-	return &DB{DB: db}, nil
+	result := &DB{DB: db}
+	if err = result.Init(ctx); err != nil {
+		return nil, fmt.Errorf("initialize database: %w", err)
+	}
+
+	return result, nil
 }
 
 // Init initializes the database schema.
 func (db *DB) Init(ctx context.Context) error {
-	schema := `
-		CREATE TABLE IF NOT EXISTS messages (
-			id INTEGER PRIMARY KEY AUTOINCREMENT,
-			chat_id INTEGER NOT NULL,
-			user_id INTEGER NOT NULL,
-			text TEXT NOT NULL,
-			created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-		);
-		CREATE INDEX IF NOT EXISTS idx_messages_chat_id ON messages(chat_id);
-		CREATE INDEX IF NOT EXISTS idx_messages_user_id ON messages(user_id);
-	`
-
-	if _, err := db.ExecContext(ctx, schema); err != nil {
-		return fmt.Errorf("create schema: %w", err)
+	if _, err := db.ExecContext(ctx, initSQL); err != nil {
+		return fmt.Errorf("create schema error: %w", err)
 	}
 
 	return nil
 }
 
-// Message represents a stored message.
-type Message struct {
-	ID        int64  `db:"id"`
-	ChatID    int64  `db:"chat_id"`
-	UserID    int64  `db:"user_id"`
-	Text      string `db:"text"`
-	CreatedAt string `db:"created_at"`
+// Event represents a load event with a timestamp and load percentage.
+type Event struct {
+	Timestamp time.Time `db:"timestamp"`
+	Load      uint8     `db:"load"`
 }
 
-// SaveMessage stores a message in the database.
-func (db *DB) SaveMessage(ctx context.Context, chatID, userID int64, text string) error {
-	query := `INSERT INTO messages (chat_id, user_id, text) VALUES (?, ?, ?)`
-	if _, err := db.ExecContext(ctx, query, chatID, userID, text); err != nil {
-		return fmt.Errorf("insert message: %w", err)
+// LogValue implements slog.LogValuer for Event.
+func (e Event) LogValue() slog.Value {
+	return slog.StringValue(fmt.Sprintf("{timestamp: '%s', load: %d}", e.Timestamp.Format(time.RFC3339), e.Load))
+}
+
+// SaveEvent stores an event in the database.
+func (db *DB) SaveEvent(ctx context.Context, event Event) error {
+	const query = `INSERT INTO events (timestamp, load) VALUES (:timestamp, :load);`
+
+	if _, err := db.NamedExecContext(ctx, query, event); err != nil {
+		return fmt.Errorf("insert event: %w", err)
 	}
+
 	return nil
 }
 
-// GetMessagesByChat retrieves messages for a specific chat.
-func (db *DB) GetMessagesByChat(ctx context.Context, chatID int64, limit int) ([]Message, error) {
-	var messages []Message
-	query := `SELECT id, chat_id, user_id, text, created_at FROM messages WHERE chat_id = ? ORDER BY created_at DESC LIMIT ?`
-	if err := db.SelectContext(ctx, &messages, query, chatID, limit); err != nil {
-		return nil, fmt.Errorf("select messages: %w", err)
+// SaveManyEvents stores multiple events in the database.
+func (db *DB) SaveManyEvents(ctx context.Context, events []Event) error {
+	if len(events) == 0 {
+		return nil
 	}
-	return messages, nil
+
+	const query = `INSERT OR REPLACE INTO events (timestamp, load) VALUES (:timestamp, :load);`
+
+	if _, err := db.NamedExecContext(ctx, query, events); err != nil {
+		return fmt.Errorf("insert events: %w", err)
+	}
+
+	return nil
+}
+
+// GetEvents retrieves events to the current time minus the given period.
+func (db *DB) GetEvents(ctx context.Context, period time.Duration, location *time.Location) ([]Event, error) {
+	const query = `SELECT timestamp, load FROM events WHERE timestamp >= ? ORDER BY timestamp;`
+	var (
+		ts     = time.Now().UTC().Add(-period)
+		events []Event
+	)
+
+	slog.DebugContext(ctx, "GetEvents", "query", query, "since", ts)
+	if err := db.SelectContext(ctx, &events, query, ts); err != nil {
+		return nil, fmt.Errorf("failed select events: %w", err)
+	}
+
+	n := len(events)
+	if n > 0 {
+		slog.DebugContext(ctx, "GetEvents", "first", events[0], "last", events[n-1], "count", n)
+	}
+
+	for _, event := range events {
+		event.Timestamp = event.Timestamp.In(location)
+	}
+
+	return events, nil
 }
 
 // Close closes the database connection.
 func (db *DB) Close() error {
 	return db.DB.Close()
+}
+
+// InTransaction executes the given function within a database transaction.
+func InTransaction(ctx context.Context, db *DB, f func(tx *sqlx.Tx) error) error {
+	tx, err := db.BeginTxx(ctx, &sql.TxOptions{Isolation: sql.LevelSerializable})
+	if err != nil {
+		return fmt.Errorf("begin transaction: %w", err)
+	}
+
+	if err = f(tx); err != nil {
+		err = fmt.Errorf("transaction function error: %w", err)
+		if rbErr := tx.Rollback(); rbErr != nil {
+			err = errors.Join(err, fmt.Errorf("rollback error: %w", rbErr))
+		}
+		return err
+	}
+
+	if err = tx.Commit(); err != nil {
+		return fmt.Errorf("commit transaction: %w", err)
+	}
+
+	return nil
+}
+
+// NewEventFromCSVRecord creates an Event from a CSV record.
+func NewEventFromCSVRecord(record []string, location *time.Location) (*Event, error) {
+	if len(record) < 2 {
+		return nil, fmt.Errorf("invalid record length: %d", len(record))
+	}
+
+	timestamp, err := time.ParseInLocation(time.DateTime, record[0], location)
+	if err != nil {
+		return nil, fmt.Errorf("parse timestamp %q: %w", record[0], err)
+	}
+
+	load, err := strconv.ParseUint(record[1], 10, 8)
+	if err != nil {
+		return nil, fmt.Errorf("parse load %q: %w", record[1], err)
+	}
+
+	return &Event{Timestamp: timestamp, Load: uint8(load)}, nil
+}
+
+// SaveManyEventsTx stores multiple events in the database within a transaction.
+func SaveManyEventsTx(ctx context.Context, tx *sqlx.Tx, events []*Event) error {
+	if len(events) == 0 {
+		return nil
+	}
+
+	const query = `INSERT OR REPLACE INTO events (timestamp, load) VALUES (:timestamp, :load);`
+
+	if _, err := tx.NamedExecContext(ctx, query, events); err != nil {
+		return fmt.Errorf("insert events: %w", err)
+	}
+
+	return nil
 }
